channels/im_api/qq/requirements.txt
aiohttp>=3.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0
